pull of updated master after merging JK-lab into mpull of updated master after merging JK-lab into master
New integration tests added for DARC. See previous commit from JK-lab for details

----
An additional 43 commit messages were squashed into this commit:
--
changes to DARC:Ray files can now be generated using a pocket grid centered at residue or a bound ligand. Default weights for the DARC parameters were changed, new weights were found by the iterative weight fitting approach. DARC default output file names were changed. Added new integration tests and removed old tests
--
changed weights for DARC parameters
--
solved conflicts after merge of master into JK-lab
--
modified DARC code to resize the espGrid to match pocketGrid, both grids now can have different grid_centers
--
added new integration tests for DARC, removed old integration tests
--
Removed prototype of undefined function copy_PocketGrid_dim_and_mid_points() in protocols/pockets/PocketGrid.hh.
--
Redefined prototype of set_reference_pose() in src/protocols/simple_moves/SuperimposeMover.hh.
--
Merged master into JK-lab.
--
Solved conflicts after merge of master into JK-lab.
--
added andrea's changes to avoid warnings
--
hardcoded addition of hydrogens for DARC electrostatics calculations and made optional for shapeonly calculations
--
This commit removes a library from the Mac clang opencl build that caused linker errors.
--
Fixed potential issue for non mac, clang opencl builds
--
settings update for MAC 10.10
--
For GPU-DARC, save eletrostatics grid as float instead of float4
--
included DARC component scores, renamed sasa calculator
--
Fixed error in separating CPU and GPU darc calc
--
Fixed error in separating CPU and GPU code for timing DARC calculations
--
Further separated CPU and GPU code to timing DARC calculations
--
added darc_shape_only option flag to ignore resizing electrostatics grid while making rays
--
Fixed a bug in checking whether ligand is out of grid; added function to copy PocketGrid dimensions
--
Fixed code issuing variable-set-but-not-used warnings in get_pharmacophore_without_bound_rna.cc.
--
Fixed warning-issuing code in src/protocols/PocketGrid.cc.
--
Fixed unused-variable warning in protocols/pockets/Fingerprint.cc.
--
Added missing try--catch blocks to apps. Fixed unused-variable warning.
--
Added missing try--catch blocks to JK-lab apps.
--
Merge between my recent changes and Ragul's recent changes in JK-lab.
--
Fixed (in a different way) "src/core/scoring/TMscore.cc" to allow linux PyRosetta to build.
--
new integration test added for DARC2, old darc integration test was made compatible to the code changes
--
separated GPU and CPU code within DARCPSO, added unique residue names for printing eggshell pdb
--
Fixed line in "src/core/scoring/TMscore.cc" that might break linux PyRosetta building.
--
Fixed problems in PocketExemplarMultifunc.cc and PocketExemplarMultifunc.hh.
--
Fixed error- and warning-issuing code in Fingerprint.hh, Fingerprint.cc, and PocketGrid.cc.
--
optimized GPU code with one-time ligand initialization, indexed origins for each ray, included multiple-origin in DARC PSO, removed MultiFingerprintMultifunc
--
Modified PocketGrid.cc and FingerPrint.hh to allow PyRosetta building.
--
Modified protocols/pockets/PocketExemplarMultifunc.hh to allow header-only building.
--
Reverted change to "if" condition in PocketGrid.cc.
--
Final fixing of warning-issuing code after JK-lab and master were merged.

Building was completed smoothly on the KU cluster (Red Hat 3.4.6-8 GCC)
--
Fixing unused variable warning in JK-lab code.
--
Fixing warning-issuing code in JK-lab after it was merged with master.
--
Further fixing of warning-issuing code in master and JK-lab after they were merged.
--
Further fixing of warning-issuing code from master and JK-lab after they were merged.
--
Fixing further warning-issuing code from master after it was merged with JK-lab.
--

Old repository SHA1: 1e0e3a1462814b2fa93580e9e1b1f7ab81736efc

+++ source/src/apps/benchmark/performance/Docking.bench.hh
@@ -54,7 +54,7 @@ public:
 		//for(int i=0; i<10; i++) {
 		//	std::cout << "i="<< i << " R=" << numeric::random::uniform() << std::endl;
 		//}
-		int reps( TScale*scaleFactor );
+		int reps( (int)(TScale*scaleFactor) );
 		if( reps == 0 ) { reps = 1; } // do at least one repetition, regardless of scaling
 		for(int i=0; i<reps; i++) {
 			core::pose::Pose pose;
